move create group from shared file
