+++ lib/storage/client.js
@@ -11,6 +11,9 @@ const {
   ClientInfo,
   EAKInfo,
   FileMeta,
+  Group,
+  GroupData,
+  GroupKeys,
   IncomingSharingPolicy,
   Meta,
   OutgoingSharingPolicy,
@@ -1045,6 +1048,19 @@ class Client extends CryptoConsumer {
     return shared.issueNoteChallenge(this, params, meta)
   }
 
+  async createGroupRequest(name, capabilities, encryptionKeyPair) {
+    const groupKeyPair = await this.crypto.generateKeypair()
+    // should this go back to using encryptAk ? maybe so.
+    const eak = await this.crypto.encryptPrivateKey(
+      groupKeyPair.privateKey,
+      encryptionKeyPair
+    )
+    let groupData = new GroupData(name, capabilities)
+    let groupKeys = new GroupKeys(encryptionKeyPair.publicKey, eak)
+    let group = new Group(groupData, groupKeys)
+    return group
+  }
+
   /**
    * createGroup creates a group and sends it to TozStore.
    *
@@ -1054,19 +1070,30 @@ class Client extends CryptoConsumer {
    * @returns {Group} A response from TozStore; the group that has been written.
    */
   async createGroup(name, capabilities = []) {
-    /* eslint-enable */
-    const encryptionKeys = new KeyPair(
+    // /* eslint-enable */
+    const encryptionKeyPair = new KeyPair(
       this.config.publicKey,
       this.config.privateKey
     )
-    return shared.createGroup(
-      this.crypto,
-      this.authenticator,
+    let createGroupReq = await this.createGroupRequest(
       name,
       capabilities,
-      encryptionKeys,
-      this.config.apiUrl
+      encryptionKeyPair
     )
+    let response = await this.authenticator.tsv1Fetch(
+      `${this.config.apiUrl}/v2/storage/groups`,
+      {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        body: createGroupReq.stringify(),
+      }
+    )
+    let storedGroupResp = await checkStatus(response)
+    let groupJson = await storedGroupResp.json()
+    let group = Group.decode(groupJson)
+    return group
   }
 }
