+++ lib/storage/shared.js
@@ -7,9 +7,6 @@ const {
   NoteInfo,
   NoteKeys,
 } = require('../../types')
-const GroupData = require('../../types/groupData')
-const GroupKeys = require('../../types/groupKeys')
-const Group = require('../../types/group')
 
 /**
  * Retrieve an access key from the server.
@@ -490,53 +487,6 @@ async function issueNoteChallenge(client, requestParams, body = {}) {
   return request.json()
 }
 
-async function createGroupRequest(
-  crypto,
-  name,
-  capabilities,
-  encryptionKeyPair
-) {
-  const groupKeyPair = await crypto.generateKeypair()
-  const eak = await crypto.encryptPrivateKey(
-    groupKeyPair.privateKey,
-    encryptionKeyPair
-  )
-  let groupData = new GroupData(name, capabilities)
-  let groupKeys = new GroupKeys(encryptionKeyPair.publicKey, eak)
-  let group = new Group(groupData, groupKeys)
-  return group
-}
-
-/*
- * createGroup is an internal method used to create a group
- */
-async function createGroup(
-  crypto,
-  authenticator,
-  name,
-  capabilities,
-  encryptionKeyPair,
-  apiUrl
-) {
-  let createGroupReq = await createGroupRequest(
-    crypto,
-    name,
-    capabilities,
-    encryptionKeyPair
-  )
-  let response = await authenticator.tsv1Fetch(`${apiUrl}/v2/storage/groups`, {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-    },
-    body: createGroupReq.stringify(),
-  })
-  let storedGroupResp = await checkStatus(response)
-  let groupJson = await storedGroupResp.json()
-  let group = Group.decode(groupJson)
-  return group
-}
-
 module.exports = {
   getEncryptedAccessKey,
   getAccessKey,
@@ -553,6 +503,4 @@ module.exports = {
   readNote,
   deleteNote,
   issueNoteChallenge,
-  createGroupRequest,
-  createGroup,
 }
